UI 구조 리팩토링 및 데이터 검증 강화
- MainApp을 WidgetRawDataProcessing과 WidgetResultsAnalyzer로 분리하여 모듈화
- DataValidator 도입: 필수 컬럼(Time, RigidBody) 및 데이터 행 수 검증 로직 추가
- PipelineController에 에러 처리 시그널(analysis_failed) 추가
- 설계 문서(design_proposal) 최신화 및 레거시 테스트 코드 정리
- 신규 단위 테스트(Validator, RawData, ResultsAnalyzer) 작성
