+++ design_proposal/component_specs.txt
@@ -8,14 +8,15 @@
 ### 1. `MainApp` (GUI 애플리케이션)
 ---
 *   **책임:**
-    *   **Original Data Analysis:** 메인 GUI 창과 원본 데이터 분석에 관련된 모든 위젯을 관리하고, 사용자 상호작용을 처리합니다. 데이터 선택을 위한 팝업 대화상자를 생성하고 관리합니다.
-    *   **Result Analyzer:** 이전에 분석된 결과 파일을 로드하고, 사용자가 그 안의 데이터를 선택하여 시각화할 수 있는 '결과 분석기' UI 전체를 관리합니다.
-*   **핵심 위젯 (Original Data Analysis):**
+    *   **Raw Data Processing:** `WidgetRawDataProcessing`을 인스턴스화하여 원본 데이터 분석 UI를 구성합니다.
+    *   **Results Analyzer:** `WidgetResultsAnalyzer`를 인스턴스화하여 결과 분석 및 시각화 UI를 구성합니다.
+    *   **Signal Handling:** 각 위젯에서 발생하는 시그널(`analysis_requested`, `analysis_failed` 등)을 처리하고 조정합니다.
+*   **핵심 위젯 (WidgetRawDataProcessing):**
     *   `Matplotlib Canvas/Toolbar`: 원본 데이터 그래프 표시 및 제어.
     *   `QPushButton` ("Load CSV File...", "Select Data...", "Run Analysis", "Export Results"): 핵심 기능 실행 버튼.
     *   `QCheckBox` ("Enable Slicing"): 분석 구간 설정을 활성화/비활성화.
     *   `QGroupBox`: 연관된 위젯들을 시각적으로 그룹화.
-*   **핵심 위젯 (Result Analyzer):**
+*   **핵심 위젯 (WidgetResultsAnalyzer):**
     *   두 번째 `Matplotlib Canvas/Toolbar`: 결과 비교용 그래프 표시.
     *   `QListWidget`: 폴더 내 결과 파일 목록 표시.
     *   `QTreeWidget`: 결과 파일 내 데이터 컬럼을 계층적으로 표시하고 선택 기능 제공.
