+++ design_proposal/flow_chart.txt
@@ -18,7 +18,9 @@
 [ 3. 사용자가 CSV 파일을 선택하고 확인 ]
    |
    +-----> [ 프로그램이 전체 CSV 파일을 메모리로 읽어들임 (RawData DataFrame 생성) ]
-   |       [ 필수 컬럼('Time', 'Position X/Y/Z' 등) 유효성 검사 ]
+   |       [ **데이터 검증 (DataValidator):** ]
+   |       [ 1. 데이터 행 수 확인 (min_rows=50) ]
+   |       [ 2. 필수 컬럼('Time', 'RigidBody_Position_X/Y/Z') 존재 여부 확인 ]
    |
    +-----> [ 만약 유효성 검사 실패 시: 사용자에게 에러 메시지 표시 후 2번 단계로 복귀 ]
    |
