+++ design_proposal/gui_description.md
@@ -8,8 +8,8 @@
 
 ## 2. 전체 레이아웃
 애플리케이션의 창은 크게 두 개의 주요 영역으로 나뉩니다:
-1.  **Original Data Analysis:** 원본 CSV 데이터를 불러와 분석을 실행하고 결과를 내보내는 주 작업 영역입니다.
-2.  **Result Analyzer:** 이미 분석이 완료된 결과 CSV 파일들을 불러와 시각화하고 비교하는 보조 작업 영역입니다.
+1.  **Raw Data Processing:** 원본 CSV 데이터를 불러와 분석을 실행하고 결과를 내보내는 주 작업 영역입니다.
+2.  **Results Analyzer:** 이미 분석이 완료된 결과 CSV 파일들을 불러와 시각화하고 비교하는 보조 작업 영역입니다.
 
 창의 가장 하단에는 **상태 표시줄(Status Bar)**이 위치합니다.
 
@@ -17,7 +17,7 @@
 
 ## 3. 세부 구성 요소
 
-### 3.1. Original Data Analysis 영역
+### 3.1. Raw Data Processing 영역
 이 영역은 원본 데이터를 처리하고 분석하는 모든 기능을 포함하며, 상단 패널과 하단 컨트롤 패널로 구성됩니다.
 
 #### 3.1.1. 상단 패널 (좌: 그래프, 우: 컨트롤)
@@ -50,7 +50,7 @@
 
 ---
 
-### 3.2. Result Analyzer 영역
+### 3.2. Results Analyzer 영역
 이 영역은 이전에 분석되어 저장된 결과 파일들을 시각화하는 데 사용됩니다.
 
 - **결과 그래프 (Result Plot Area):**
