+++ design_proposal/software_design_document.md
@@ -12,11 +12,11 @@
 
 ### 2.1. 주요 기능 영역
 
-- **원본 데이터 분석 (Original Data Analysis):**
+- **원시 데이터 처리 (Raw Data Processing):**
     - 원본 CSV 파일을 로드하고, 분석 옵션(예: 슬라이싱 범위)을 설정하며, 분석 파이프라인을 실행하고, 최종 결과를 내보내는 핵심 기능 영역입니다.
     - 주요 상호작용으로는 Matplotlib 그래프를 통한 데이터 시각화, 데이터 선택 대화상자, 상호작용형 슬라이싱 등이 포함됩니다.
 
-- **결과 분석기 (Result Analyzer):**
+- **결과 분석기 (Results Analyzer):**
     - 이전에 분석되어 CSV 파일로 저장된 결과들을 불러와 시각화하고 비교 분석하는 보조 기능 영역입니다.
     - 사용자는 결과 파일 목록에서 파일을 선택하고, `QTreeWidget`에 표시된 데이터 컬럼들 중 원하는 항목을 선택하여 두 번째 Matplotlib 그래프에 플로팅할 수 있습니다.
     - 그래프의 특정 지점을 클릭하거나, 특정 데이터 시리즈의 최대값을 찾는 '지점 분석(Point Analysis)' 기능을 통해 상세 데이터를 확인하고 내보낼 수 있습니다.
@@ -40,7 +40,9 @@
 
 ### 4.1. GUI 컴포넌트
 
-- **`MainApp` (GUI):** 사용자 입력을 받고, '원본 데이터 분석' 및 '결과 분석기'의 모든 UI 이벤트를 처리하며, 분석 실행을 `PipelineController`에 위임합니다.
+- **`MainApp` (GUI):** 사용자 입력을 받고, '원시 데이터 처리' 및 '결과 분석기'의 모든 UI 이벤트를 처리하며, 분석 실행을 `PipelineController`에 위임합니다.
+- **`WidgetRawDataProcessing`:** 원본 데이터 로드, 설정, 분석 실행 UI를 담당하는 독립 위젯입니다.
+- **`WidgetResultsAnalyzer`:** 결과 파일 로드, 시각화, 시나리오 출력을 담당하는 독립 위젯입니다.
 - **`DataSelectionDialog` (대화상자):** 다중 데이터 선택 UI를 제공합니다.
 - **`PlotManager` (시각화):** Matplotlib 캔버스를 관리하고, 상호작용형 그래프(줌, 팬, 슬라이싱 등)를 처리합니다.
 
@@ -50,6 +52,8 @@
     - **책임:** 전체 분석 파이프라인의 **조립 및 실행**을 담당하는 오케스트레이터 역할을 합니다.
     - GUI로부터 받은 설정값을 바탕으로 필요한 `Analysis Modules`를 초기화하고, 정의된 순서대로 모듈들을 순차적으로 호출합니다.
     - 실제 계산 로직을 포함하지 않으며, 각 분석 단계의 실행 흐름만 제어합니다.
+    - **에러 처리:** 분석 중 발생하는 예외를 포착하여 `analysis_failed` 시그널을 통해 GUI에 알립니다.
+    - **데이터 검증:** `DataValidator`를 사용하여 분석 시작 전 데이터의 유효성을 검사합니다.
 
 - **`Analysis Modules` (분석 기능 단위):**
     - **책임:** 각각의 독립적인 분석 기능을 수행합니다.
